@@ -17,15 +17,14 @@ Troll::Troll(ngl::Vec3 _pos,ngl::Vec3 _dir,ngl::Vec4 _c,const std::string &_shad
                              _parent
 														)
 {
-	ngl::Random *rng=ngl::Random::instance();
 
-	m_rotX=rng->randomNumber(360);
-	m_rotY=rng->randomNumber(360);
-	m_rotZ=rng->randomNumber(360);
+	m_rotX=ngl::Random::randomNumber(360);
+	m_rotY=ngl::Random::randomNumber(360);
+	m_rotZ=ngl::Random::randomNumber(360);
 
-	m_rotUpdateX=rng->randomNumber(10);
-	m_rotUpdateY=rng->randomNumber(10);
-	m_rotUpdateZ=rng->randomNumber(10);
+	m_rotUpdateX=ngl::Random::randomNumber(10);
+	m_rotUpdateY=ngl::Random::randomNumber(10);
+	m_rotUpdateZ=ngl::Random::randomNumber(10);
 
 }
 
@@ -33,9 +32,8 @@ Troll::Troll(ngl::Vec3 _pos,ngl::Vec3 _dir,ngl::Vec4 _c,const std::string &_shad
 void Troll::draw() const
 {
 
-  ngl::ShaderLib *shader=ngl::ShaderLib::instance();
-  (*shader)[m_shaderName]->use();
-  shader->setUniform("Colour",m_colour);
+  ngl::ShaderLib::use(m_shaderName);
+  ngl::ShaderLib::setUniform("Colour",m_colour);
 
   ngl::Transformation t;
   t.setPosition(m_pos);
@@ -44,20 +42,15 @@ void Troll::draw() const
                 m_parent->getCamera()->view *
                 m_parent->getGlobalTransform() *
                 t.getMatrix();
-  shader->setUniform("MVP",MVP);
+  ngl::ShaderLib::setUniform("MVP",MVP);
 
-    // get an instance of the VBO primitives for drawing
-    ngl::VAOPrimitives *prim=ngl::VAOPrimitives::instance();
-    // draw the sphere
-    prim->draw("troll");
+  ngl::VAOPrimitives::draw("troll");
 
 }
 
 
 void Troll::update()
 {
-// more the particle by adding the Direction vector to the position
-	ngl::Random *rng=ngl::Random::instance();
 
   m_pos+=m_dir;
   ++m_life;
@@ -70,14 +63,14 @@ void Troll::update()
     m_life=0.0;
     m_pos=m_parent->getPos();
 
-    m_maxLife=rng->randomPositiveNumber(80);
+    m_maxLife=ngl::Random::randomPositiveNumber(80);
 
-		m_rotX=rng->randomNumber(360);
-		m_rotUpdateX=rng->randomPositiveNumber(10);
-		m_rotY=rng->randomNumber(360);
-		m_rotUpdateY=rng->randomPositiveNumber(10);
-		m_rotZ=rng->randomNumber(360);
-		m_rotUpdateZ=rng->randomPositiveNumber(10);
+		m_rotX=ngl::Random::randomNumber(360);
+		m_rotUpdateX=ngl::Random::randomPositiveNumber(10);
+		m_rotY=ngl::Random::randomNumber(360);
+		m_rotUpdateY=ngl::Random::randomPositiveNumber(10);
+		m_rotZ=ngl::Random::randomNumber(360);
+		m_rotUpdateZ=ngl::Random::randomPositiveNumber(10);
 	}
 }
 
